requirements.txt
###### Requirements ######
struct
glob
pyelftools==0.27
